requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
Pillow
requests
